Split components in the checkout page
