+++ src/components/checkout/customerForm/index.js
@@ -1,0 +1,3 @@
+import CustomerForm from './customerForm';
+
+export default CustomerForm;
