+++ src/components/checkout/orderSummary/index.js
@@ -1,0 +1,3 @@
+import OrderSummary from './orderSummary';
+
+export default OrderSummary;
